facesyma_backend/requirements.txt
django==4.2.16
djangorestframework==3.15.2
django-cors-headers==4.4.0
pymongo==4.7.2
PyJWT==2.9.0
google-auth==2.35.0
gunicorn==23.0.0
daphne==4.1.2
channels==4.1.0
channels-redis==4.1.0
Pillow==10.4.0
numpy==1.26.4
opencv-python-headless==4.10.0.84
mediapipe==0.10.14
openpyxl==3.1.2
pandas==2.2.2
tensorflow==2.17.0
apscheduler==3.10.4
stripe==10.12.0
